requirements.txt
streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
opencv-python-headless==4.8.1.78
numpy==1.23.5
albumentations==1.3.1
Pillow==10.2.0
tensorflow-cpu==2.13.0
matplotlib==3.8.2 